+++ pages/settings/modules/[module]/edit.vue
@@ -7,131 +7,11 @@
       {{ generalError.main }}
     </div>
     <div class="flex flex-col gap-4 w-full">
-      <div class="separator" />
-      <h2 class="font-bold text-base-plus">
-        Module
-      </h2>
-      <div class="grid gap-4 grid-cols-1 md:grid-cols-2 lg:grid-cols-4">
-        <UiTextInput
-          v-model="form.name"
-          :errors="formErrors.data?.name"
-          label="Module name"
-          name="module-name"
-          @update:model-value="setModuleKey"
-        />
-        <UiTextInput
-          v-model="form.key"
-          :errors="formErrors.data?.key"
-          label="Module key"
-          name="module-key"
-          disabled
-        />
-      </div>
-      <div class="separator" />
-      <h2 class="font-bold text-base-plus">
-        Fields
-      </h2>
-      <ul
-        ref="fieldList"
-        class="grid gap-4 border-solid border-primary rounded-lg"
-        :class="{'border p-4': form.fields.length}"
-      >
-        <li v-for="(field, index) in form.fields" :key="index" class="grid gap-4">
-          <div class="flex gap-4 w-full">
-            <UiButton
-              class="w-max text-sm"
-              error-variant
-              @click.prevent="deleteField(index)"
-            >
-              Remove field
-            </UiButton>
-            <div v-if="generalError && field.key === ''" class="text-error-border text-left">
-              {{ generalError.field }}
-            </div>
-          </div>
-          <div class="grid gap-4 grid-cols-1 md:grid-cols-2 lg:grid-cols-4">
-            <UiTextInput
-              v-model="field.title"
-              label="Field name"
-              :name="`field-name-${index}`"
-              :errors="formErrors.data?.fields[field.key]?.title"
-              @update:model-value="setFieldKey(index)"
-            />
-            <UiTextInput
-              v-model="field.key"
-              label="Field key"
-              :name="`module-key-${index}`"
-              :errors="formErrors.data?.fields[field.key]?.key"
-              disabled
-            />
-            <!-- @vue-ignore - due to field.valueType = undefined -->
-            <UiSelect
-              v-model="field.type"
-              :items="fieldTypeItems"
-              :name="`field-type-${index}`"
-              label="Field type"
-              :errors="formErrors.data?.fields[field.key]?.type"
-              @update:model-value="field.valueType = undefined; form.fields[index].additional = {}; getFieldValueItems(field.type, index)"
-            />
-            <UiSelect
-              :ref="`fieldValue${index}`"
-              v-model="field.valueType"
-              :items="getFieldValueItems(field.type, index)"
-              :name="`field-value-type-${index}`"
-              label="Field value type"
-              :disabled="!field.type || fieldValueTypeMap[field.type].length === 1"
-              :hide-cross="!field.type || fieldValueTypeMap[field.type].length === 1"
-              :errors="formErrors.data?.fields[field.key]?.valueType"
-            />
-            <div v-if="getAdditionalFieldType(field.type, field.valueType)?.name === 'arrayValueType'" class="flex flex-col gap-2">
-              <UiSelect
-                v-model="field.additional.arrayValueType"
-                :items="fieldValueItemsArrayType"
-                :name="`field-addditionals-${field.type}-${field.valueType}-${index}`"
-                :label="getAdditionalFieldType(field.type, field.valueType)?.inputLabel"
-              />
-              <UiCheckbox
-                v-model="field.additional.multiselect"
-                :name="`field-addditionals-${field.type}-${field.valueType}-multiselect-${index}`"
-                label="Multiselect?"
-              />
-            </div>
-            <div v-if="field.type === 'fileUpload' || field.type === 'imageUpload'" class="flex gap-4">
-              <UiTextInput
-                v-model="field.additional.maxFileSizeMb"
-                :name="`field-addditionals-${field.type}-${field.valueType}-max-size-${index}`"
-                :label="getAdditionalFieldType(field.type, field.valueType)?.inputLabel"
-                type="number"
-              />
-              <UiTextInput
-                v-model="field.additional.buttonTitle"
-                :name="`field-addditionals-${field.type}-${field.valueType}-button-title-${index}`"
-                label="Button title"
-              />
-              <UiCheckbox
-                v-model="field.additional.multipleFiles"
-                :name="`field-addditionals-${field.type}-${field.valueType}-multiple-${index}`"
-                label="Multiple?"
-              />
-            </div>
-
-            <!-- @vue-ignore -->
-            <UiTextInput
-              v-else-if="
-                getAdditionalFieldType(field.type, field.valueType)
-                  &&
-                  getAdditionalFieldType(field.type, field.valueType)!.name !== 'multiselect'
-              "
-              v-model="field.additional[getAdditionalFieldType(field.type, field.valueType)!.name]"
-              :name="`field-addditionals-${field.type}-${field.valueType}-${index}`"
-              :label="getAdditionalFieldType(field.type, field.valueType)?.inputLabel"
-            />
-          </div>
-          <UiCheckbox v-model="field.required" label="Required" :name="`field-required-${index}`" />
-          <div class="separator" />
-        </li>
-      </ul>
-      <UiButton text="Add field" secondary class="w-full md:w-[300px]" @click="addField" />
+      <ModuleSettingsForm
+        v-model="form"
+        :errors="formErrors"
+        :general-error="generalError"
+      />
       <div class="separator" />
       <div v-if="!loading" class="flex justify-between">
         <UiButton
@@ -166,8 +46,6 @@
 </template>
 
 <script setup lang="ts">
-import { EFieldType } from '@prisma/client';
-import { UiSelect, UiTextInput } from '#components';
 import type { ModuleFieldsAdjusted, ResponseError } from '~/types';
 
 defineOptions({
@@ -189,61 +67,12 @@ const form = ref({
     fields: [] as ModuleFieldsAdjusted[]
 });
 
-const fieldList = ref();
 const formErrors = ref<ResponseError>({});
-const generalError = ref();
+const generalError = ref<{
+  main: string,
+  field: string
+}>();
 const loading = ref(false);
-
-function getFieldValueItems (fieldType: EFieldType, index: number) {
-    if (!fieldType) {
-        return [];
-    }
-    const currentArray = fieldValueTypeMap[fieldType];
-
-    if (currentArray.length === 1) {
-        form.value.fields[index].valueType = currentArray[0];
-    }
-
-    return currentArray.map((value) => {
-        return {
-            key: value,
-            title: value.toUpperCase(),
-            position: 0,
-            visible: true
-        };
-    });
-}
-
-const fieldValueItemsArrayType = [
-    {
-        key: 'string',
-        title: 'STRING',
-        position: 0,
-        visible: true
-    },
-    {
-        key: 'number',
-        title: 'NUMBER',
-        position: 1,
-        visible: true
-    }
-];
-
-function addField () {
-    form.value.fields.push({ ...fieldTemplate });
-}
-
-function deleteField (index: number) {
-    form.value.fields.splice(index, 1);
-}
-
-function setModuleKey () {
-    form.value.key = sanitizeTitleToKey(form.value.name);
-}
-
-function setFieldKey (index: number) {
-    form.value.fields[index].key = sanitizeTitleToKey(form.value.fields[index].title);
-}
 
 async function getModuleWithFields () {
     const { data } = await $fetch('/api/data/modules/getOne', {
@@ -494,7 +323,3 @@ onMounted(async () => {
 });
 
 </script>
-
-<style scoped>
-
-</style>
